Adds custom CA Lokup for Linux
Implements search of custom CA certificates in linux. The approach is based on OpenSSL directories and the fact that most Linux distros use the `OPENSSLDIR` as the repo for certificates.
THe documented process for installing certificates ([Ubuntu /etc/ssl/](https://help.ubuntu.com/lts/serverguide/certificates-and-security.html#installing-the-certificate), [Ubuntu, Debian update-ca-certificates](http://askubuntu.com/a/649463), [RedHat, CentOS, Fedora update-ca-trust](https://www.happyassassin.net/2015/01/14/trusting-additional-cas-in-fedora-rhel-centos-dont-append-to-etcpkitlscertsca-bundle-crt-or-etcpkitlscert-pem/)) will make custom CA authorities available for the connector.